Check load-languages in troubleshooting guide
See https://github.com/astahlman/ob-async/issues/21, in which the
subprocess was not able to find the function `org-babel-execute:sql`
because `sql` had never been added to `org-babel-load-languages` (in the
parent process).

I've added a new checklist entry in troubleshooting.org to help diagnose
this type of problem by printing the list of languages defined in
`org-babel-load-languages`.