ob-async: use with-current-buffer to update org buffer
Currently when updating the results buffer the user might get
surprised if they are in the middle of typing. This is because using
switch-to-buffer will interfere with interactive typing, as the
documentation notes:

  WARNING: This is NOT the way to work on another buffer temporarily
within a Lisp program!

Rather than using set-buffer though we can simply use the
`with-current-buffer' wrapper. Although this works fine for updating
the another buffer we also need to do a save-excursion so if the user
is using this buffer their position is not lost. The user will still
see the "Code block evaluation complete." message appear in the
miibuffer but we could make a cleaner message. This is left as an
exercise for a later patch.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>