Add list of languages for ob-async to ignore (#35)
I recently noticed an incompatibility between this and the ob-ipython package, which defines its own :async keyword (and keeps an open pipe to an ipython session running in the background). To remedy this, I have added an elisp variable ob-async-no-async-languages-alist here: for any languages included in this list, the original ctrl-c-ctrl-c function is run — as it is for src blocks that do not include the :async keyword. By default, the variable is set to nil, so that there are no changes to the default behavior of ob-ipython.

I've also included a test to show this functionality in action. By setting ob-async-no-async-languages-alist to '("sh"), we can run one of the core tests, but expect that the code will finish before the check is run, circumventing the :async keyword.

Let me know if something seems amiss, and I'm happy to update this. (...and thanks for writing this package; I use it all the time.)