Determine temp files based on the test case (#66)
We have a couple of tests that ostensibly write a string to "/tmp/foo",
but neither of them deletes the file on setup or teardown. If one of
them fails, the success of the other test could mask it.

Ideally the file would be named with a GUID and we'd clean it up after
each test case, but I'll leave that as a future improvement.