Handle signals of code-blocks by returning an error as the result
During the execution of code blocks an error could be signaled.  The
consequence of this would be that the FINISH-FUNC will not be called.
Therefore handle the error in the START-FUNC and return a textual
representation of it instead of the result.

A code block like the following

```
  #+begin_src emacs-lisp :async
  (error "somewhere an error was signaled")
  #end_src
```

would no longer lead to a #+result that gets never updated and keeps
the UUID forever.  Instead the following would be put as the result:

  Failure: during the async execution of your code block the following error was signaled:
    error: (somewhere an error was signaled)