Advice org-babel-eval-error-notify to return error details if needed
At least shell and python source code blocks could fail without any
possibility to notice it.  Therefore adding an advice to
org-babel-eval-error-notify that will set the variable
ob-async-eval-error in the case of an error.  If this variable is set,
return the error.

With this the following code blocks should produce a result that
gives a hint about the error:

```
  #+begin_src sh :results output :async
  echo something went wrong >&2
  false
  #+end_src
```

```
  #+begin_src python :async
  raise NotImplementedError
  #+end_src
```